chore: tech debt batch 4 — shared fixtures + pre-commit set -e
- #170: Add set -e to pre-commit hook (abort on lint/format failure)
- #180: Extract shared jf_client fixture to integration conftest
- #181: Extract shared library_store fixture to integration conftest
- #184: Add alice_auth fixture, replace 13 duplicate authenticate calls

Closes #170, #180, #181, #184
